TimeZoneOffset Test commentet: works only standalone (UTC ist set before an TZ is not mutable after first setting) works not under win Some tests renamed eg. PrintGMonth -> GMonth (all String based tests goes here in Future)
